+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Changelog
 ## [0.1.2] - Changelog for upcoming release
+### Added
+- NoodlePlot for collective events
+
 ### Changed
 - binarize_detrend: converted pandas operations to numpy for performance improvements
 - detect_events: converted pandas operations to numpy for performance imporovements
 - stats: converted pandas operations to numpy for performance improvements
 - various small changes
 - updated docstrings to match changes
+
+### Fixed
+- numpy warning caused by stats module
 
 ## [0.1.1] - 2022-04-04
 ### Added